refactor out draggable popover
